+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to ResumeSendtoCompany will be documented in this file.
+All notable changes to Autonomous Outreach System will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -28,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2026-08-14
 
 ### Added
-- Initial release of ResumeSendtoCompany
+- Initial release of Autonomous Outreach System
 - Autonomous lead discovery system
 - AI-powered email personalization
 - Multi-provider SMTP relay pool